UPDATE: Fixes
Both issues have been fixed. Global variables are now visually
represented on the screen and loot disappears after a delayed period of
time. One more issue has been found which is that the addition of
"money" to the global money variable is continual until the instance of
the money object has disappeared. It must only do this action once. As
well as the issue of randomizing the amount of money gained by the
player when they collide with the money object.